refactor(ekb): rename service folders with _service suffix
