+++ pipelines/enterprise_knowledge_base/document_classification/bq_service/schemas.py
