+++ pipelines/enterprise_knowledge_base/document_classification/gcs_service/schemas.py
